# Faster import part2: Delay import of heavy packages