--- a/.github/workflows/ci-code.yml
+++ b/.github/workflows/ci-code.yml
@@ -148,4 +148,5 @@ jobs:
     - name: Run verdi
       run: |
         verdi devel check-load-time
+        verdi devel check-undesired-imports
         .github/workflows/verdi.sh