--- a/aiida/manage/__init__.py
+++ b/aiida/manage/__init__.py
@@ -42,7 +42,6 @@ __all__ = (
     'Option',
     'Postgres',
     'PostgresConnectionMode',
-    'ProcessLauncher',
     'Profile',
     'RabbitmqManagementClient',
     'check_and_migrate_config',