--- a/aiida/manage/external/rmq/launcher.py
+++ b/aiida/manage/external/rmq/launcher.py
@@ -9,8 +9,6 @@ import plumpy
 
 LOGGER = logging.getLogger(__name__)
 
-__all__ = ('ProcessLauncher',)
-
 
 class ProcessLauncher(plumpy.ProcessLauncher):
     """A sub class of :class:`plumpy.ProcessLauncher` to launch a ``Process``.