@sphuber 2023-09-04
Thanks @danielhollas . The changes look good to me. You are right that the removal of the export of `ProcessLauncher` to higher modules is technically breaking, but I agree that it is unlikely that it affects anyone, so I think this is an acceptable change. Are these the last of the changes related to the import speed? Or do you have more lined up?