@danielhollas 2023-09-04
Thanks @sphuber ! I have more lined up, specifically targeting verdi.